// src/types/actor/parts/PrimaryCharacteristics.ts
import { FoundryPrimaryCharacteristicField } from "../fields/FoundryPrimaryCharacteristicField";

export type PrimaryCharacteristics = {
  agility: FoundryPrimaryCharacteristicField;
  constitution: FoundryPrimaryCharacteristicField;
  dexterity: FoundryPrimaryCharacteristicField;
  strength: FoundryPrimaryCharacteristicField;
  intelligence: FoundryPrimaryCharacteristicField;
  perception: FoundryPrimaryCharacteristicField;
  power: FoundryPrimaryCharacteristicField;
  willPower: FoundryPrimaryCharacteristicField;
};