// src/types/actor/parts/AthleticsSkills.ts
import { FoundryCalculableField } from "../fields/FoundryCalculableField";

export type AthleticsSkills = {
  acrobatics: FoundryCalculableField;
  athleticism: FoundryCalculableField;
  ride: FoundryCalculableField;
  swim: FoundryCalculableField;
  climb: FoundryCalculableField;
  jump: FoundryCalculableField;
  piloting: FoundryCalculableField;
};